python/scipy-stack/requirements.txt
numpy==1.11.3
scipy==0.18.1
pillow==4.0.0
matplotlib==1.5.3